rfct: extract the AboutCommand into its own service
