+++ core/src/main/java/dev/buildcli/core/utils/BuildCLIService.java
@@ -109,14 +109,6 @@ public class BuildCLIService {
     return validCommands != null && validCommands.contains(input);
   }
 
-  public static void about() {
-    SystemOutLogger.log("BuildCLI is a command-line interface (CLI) tool for managing and automating common tasks in Java project development.\n" +
-        "It allows you to create, compile, manage dependencies, and run Java projects directly from the terminal, simplifying the development process.\n");
-    SystemOutLogger.log("Visit the repository for more details: https://github.com/BuildCLI/BuildCLI\n");
-
-    SystemOutLogger.log(gitExec.showContributors());
-  }
-
   private static void updateBuildCLI() {
     if (updateRepository()) {
       generateBuildCLIJar();
